+++ src/components/generic/errorComponent.tsx
@@ -1,0 +1,25 @@
+import { FaTimesCircle } from 'react-icons/all';
+import React from 'react';
+
+interface ErrorComponentInterface {
+    desc?: string;
+}
+
+export const ErrorComponent = ({ desc }: ErrorComponentInterface) => {
+    return (
+        <div className="w-1/2 mx-auto space-y-6">
+            <div className="px-4 py-3 text-red-700 bg-red-100 rounded-lg flex justify-center items-center">
+                <FaTimesCircle className="mr-1" />
+                Wystąpił błąd!
+            </div>
+            <div>
+                {desc && (
+                    <details>
+                        <summary className="font-bold mb-3">Szczegóły</summary>
+                        {desc}
+                    </details>
+                )}
+            </div>
+        </div>
+    );
+};
