+++ src/components/generic/loadingComponent.tsx
@@ -1,0 +1,18 @@
+import React, { useEffect, useState } from 'react';
+import { Loading } from 'react-loading-dot';
+
+export const LoadingComponent = () => {
+    const [visible, setVisible] = useState(false);
+
+    useEffect(() => {
+        const intervalID = setTimeout(() => setVisible(true), 700);
+
+        return () => {
+            clearInterval(intervalID);
+        };
+    }, []);
+
+    return visible ? (
+        <Loading background={'rgb(79, 70, 229)'} duration={'0.7s'} margin={'0.5rem'} size={'1rem'} />
+    ) : null;
+};
